Implemented System for User to Delete Account with needing to verify before doing so

+++ Features/DeleteAccount.py
@@ -8,12 +8,13 @@ current_directory = os.path.dirname(__file__)
 database_path = os.path.join(current_directory, '..', 'Database', 'CentralisedDatabase.db')
 
 
-def deleleAccount(CustomerID):
-    connection = sqlite3.connect(database_path)
-    cursor = connection.cursor()
-    cursor.execute(f"DELETE FROM LoginInformation WHERE CustomerID = ?", (CustomerID,))
-    cursor.execute(f"DELETE FROM Customer WHERE CustomerID = ?", (CustomerID,))
-    cursor.execute(f"DELETE FROM Purchase WHERE CustomerID = ?", (CustomerID,))
+def deleleAccount(CustomerID, confirmation):
+    if confirmation:
+        connection = sqlite3.connect(database_path)
+        cursor = connection.cursor()
+        cursor.execute(f"DELETE FROM LoginInformation WHERE CustomerID = ?", (CustomerID,))
+        cursor.execute(f"DELETE FROM Customer WHERE CustomerID = ?", (CustomerID,))
+        cursor.execute(f"DELETE FROM Purchase WHERE CustomerID = ?", (CustomerID,))
 
-    connection.commit()
+        connection.commit()
